+++ editor/editor-common/src/extensibility/Extension/Lozenge.tsx
@@ -4,31 +4,23 @@ import { Component } from 'react';
 import { jsx } from '@emotion/react';
 
 import type { Node as PmNode } from '@atlaskit/editor-prosemirror/model';
-import EditorFileIcon from '@atlaskit/icon/glyph/editor/file';
 
 import { getExtensionLozengeData } from '../../utils';
 
-import {
-  placeholderFallback,
-  placeholderFallbackParams,
-  styledImage,
-} from './styles';
-
-export const capitalizeFirstLetter = (str: string): string => {
-  return str.charAt(0).toUpperCase() + str.slice(1);
-};
+import { LozengeComponent } from './LozengeComponent';
+import { styledImage } from './styles';
 
 export interface Props {
   node: PmNode;
+  showMacroInteractionDesignUpdates?: boolean;
+  isNodeSelected?: boolean;
 }
 
-interface LozengeData {
+export interface LozengeData {
   url: string;
   height?: number;
   width?: number;
 }
-
-export const ICON_SIZE = 24;
 
 export default class ExtensionLozenge extends Component<Props, any> {
   render() {
@@ -43,13 +35,14 @@ export default class ExtensionLozenge extends Component<Props, any> {
     return this.renderFallback(iconData);
   }
 
-  private renderImage(lozengeData: LozengeData) {
+  private renderImage = (lozengeData: LozengeData) => {
     const { extensionKey } = this.props.node.attrs;
     const { url, ...rest } = lozengeData;
     return <img css={styledImage} src={url} {...rest} alt={extensionKey} />;
-  }
+  };
 
-  private renderFallback(lozengeData?: LozengeData) {
+  private renderFallback = (lozengeData?: LozengeData) => {
+    const { showMacroInteractionDesignUpdates, isNodeSelected } = this.props;
     const { parameters, extensionKey } = this.props.node.attrs;
     const { name } = this.props.node.type;
     const params = parameters && parameters.macroParams;
@@ -60,26 +53,17 @@ export default class ExtensionLozenge extends Component<Props, any> {
         parameters.macroMetadata.title) ||
       extensionKey;
     const isBlockExtension = name === 'extension';
+
     return (
-      <div data-testid="lozenge-fallback" css={placeholderFallback}>
-        {lozengeData && !isBlockExtension ? (
-          this.renderImage({
-            height: ICON_SIZE,
-            width: ICON_SIZE,
-            ...lozengeData,
-          })
-        ) : (
-          <EditorFileIcon label={title} />
-        )}
-        <span className="extension-title">{capitalizeFirstLetter(title)}</span>
-        {params && !isBlockExtension && (
-          <span css={placeholderFallbackParams}>
-            {Object.keys(params).map(
-              (key) => key && ` | ${key} = ${params[key].value}`,
-            )}
-          </span>
-        )}
-      </div>
+      <LozengeComponent
+        isNodeSelected={isNodeSelected}
+        showMacroInteractionDesignUpdates={showMacroInteractionDesignUpdates}
+        isBlockExtension={isBlockExtension}
+        lozengeData={lozengeData}
+        params={params}
+        title={title}
+        renderImage={this.renderImage}
+      />
     );
-  }
+  };
 }