+++ editor/editor-plugin-editor-viewmode/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-plugin-editor-viewmode
 
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies
+
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.0.25
 
 ### Patch Changes